Create Python Exponential Population 1.0
This is the Initial Version of my Python Exponential Population. Using the compounding continuously formula, you can input your initial population, along with birth rate / 100, and the amount of time lapses for the calculator. It would print out each year with the current population. 2.0 will use python and bring a GUI element for users to use.